enhanced extras/Fletcher_compare_implementations/basic.h

+++ extras/Fletcher_compare_implementations/basic.h
@@ -8,12 +8,12 @@
 
 uint16_t basic_fletcher16(uint8_t *data, const size_t length)
 {
-  uint16_t s1 = 0;
-  uint16_t s2 = 0;
+  uint8_t s1 = 0;
+  uint8_t s2 = 0;
   for (size_t i = 0; i < length; i++)
   {
-    s1 = (s1 + data[i]) % ((((uint16_t) 1) << 8) - ((uint16_t) 1));
-    s2 = (s2 + s1) % ((((uint16_t) 1) << 8) - ((uint16_t) 1));
+    s1 = (uint8_t) (((uint16_t) (s1 + data[i])) % ((((uint16_t) 1) << 8) - ((uint16_t) 1)));
+    s2 = (uint8_t) (((uint16_t) (s2 + s1)) % ((((uint16_t) 1) << 8) - ((uint16_t) 1)));
   }
   return (((uint16_t) s2) << 8) | ((uint16_t) s1);
 }
@@ -24,8 +24,8 @@ uint32_t basic_fletcher32(uint16_t *data, const size_t length)
   uint16_t s2 = 0;
   for (size_t i = 0; i < length; i++)
   {
-    s1 = (s1 + data[i]) % ((((uint32_t) 1) << 16) - ((uint32_t) 1));
-    s2 = (s2 + s1) % ((((uint32_t) 1) << 16) - ((uint32_t) 1));
+    s1 = (uint16_t) (((uint32_t) (s1 + data[i])) % ((((uint32_t) 1) << 16) - ((uint32_t) 1)));
+    s2 = (uint16_t) (((uint32_t) (s2 + s1)) % ((((uint32_t) 1) << 16) - ((uint32_t) 1)));
   }
   return (((uint32_t) s2) << 16) | ((uint32_t) s1);
 }
@@ -36,8 +36,8 @@ uint64_t basic_fletcher64(uint32_t *data, const size_t length)
   uint32_t s2 = 0;
   for (size_t i = 0; i < length; i++)
   {
-    s1 = (s1 + data[i]) % ((((uint64_t) 1) << 32) - ((uint64_t) 1));
-    s2 = (s2 + s1) % ((((uint64_t) 1) << 32) - ((uint64_t) 1));
+    s1 = (uint32_t) (((uint64_t) (s1 + data[i])) % ((((uint64_t) 1) << 32) - ((uint64_t) 1)));
+    s2 = (uint32_t) (((uint64_t) (s2 + s1)) % ((((uint64_t) 1) << 32) - ((uint64_t) 1)));
   }
   return (((uint64_t) s2) << 32) | ((uint64_t) s1);
 }
